metasploit-mcp/pyproject.toml
[project]
name = "metasploit-mcp"
version = "0.1.0"
description = "MCP Server for Metasploit Framework"
authors = [{name = "DineshPathro", email = "dineshpathro@proton.me"}]
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "python-metasploit>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
